refactor(cli): extract ingest command family into cli/commands/ingest.ts
ingest/derive/derive-signals/derive-intents move to commands/ingest.ts.
cmdIngest/cmdIngestHere keep their string-args interface (runIngest({args})
parses downstream - documented exception); cmdDeriveSignals/cmdIngestInsights
convert to typed options and progressModeFor is deleted (Flag.choice already
yields a validated ProgressMode). resolveIngestStages/detectRemovedIngestFlag/
insightsOnlyConflicts re-exported from index.ts for the existing test
contract. withIngest progress gating stays in index.ts untouched.
LEGACY_RUNTIME shrinks to agents only.
